pyproject.toml
[project]
name = "agno-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.4.5",
    "duckduckgo-search>=8.0.1",
    "groq>=0.24.0",
    "lancedb>=0.22.0",
    "openai>=1.77.0",
    "pandas>=2.2.3",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.0",
    "tantivy>=0.24.0",
    "torch>=2.7.0",
    "yfinance>=0.2.59",
]
